refactor(discover): extract filter helpers to lib and dedupe KPOP_NAME
