+++ src/app/(user)/discover/_components/ExploreMapView.tsx
@@ -6,7 +6,8 @@ import { useToast } from "../../_hooks/useToast";
 import { dedupeEventMarkers } from "@/lib/event-utils";
 import { EVENT_RED, getDDay, sortEventMarkers } from "@/lib/event-format";
 import { useSearchParams, useRouter, usePathname } from "next/navigation";
-import { parseFilterParams, serializeFilterParams, buildTopicColorMap } from "@/lib/filter-params";
+import { parseFilterParams, serializeFilterParams, buildTopicColorMap, KPOP_NAME } from "@/lib/filter-params";
+import { postMatchesFilters, placeMatchesFilters, placeMatchScore, findTopicIdBySlug } from "@/lib/discover-filter-utils";
 import { InteractiveMap, type FocusCameraHandle } from "@/components/maps/InteractiveMap";
 import { PlaceBottomSheet } from "@/components/maps/PlaceBottomSheet";
 import { PlaceListSheet, getSheetHeight, type PlaceListSheetState } from "@/components/maps/PlaceListSheet";
@@ -43,7 +44,6 @@ import type {
 import type { CuratedSectionWithSlug, SectionData } from "@/lib/curation-types";
 
 type ChipInfo = { id: string; label: string; bg: string; fg: string };
-const KPOP_NAME = "K-POP";
 const CATEGORY_ORDER: string[] = [
   "CONCERT", "LANDMARK_LIGHTING", "PROMOTION", "ACTIVITY",
   "SHOPPING", "MOBILITY", "FNB", "STAY", "WELCOME_KIT",
@@ -62,76 +62,9 @@ function calcEventPassesFilter(
   return matchesSearch && matchesCategory && matchesSaved;
 }
 
-function postMatchesFilters(post: MapPost, topicIds: string[], tagIds: string[], tagGroupKeys: string[]): boolean {
-  const topicHit =
-    topicIds.length > 0 &&
-    post.topics.some((t) => topicIds.some((id) => topicMatchesFilter(t, id)));
-  const tagHit = tagIds.length > 0 && post.tags.some((tag) => tagIds.includes(tag.id));
-  const groupHit = tagGroupKeys.length > 0 && post.allTagGroups.some((g) => tagGroupKeys.includes(g));
-  return topicHit || tagHit || groupHit;
-}
-
-function placeMatchesFilters(
-  place: Pick<MapPlace, "id" | "posts" | "area">,
-  hasPostLevelFilter: boolean,
-  matchedPostsByPlaceId: Map<string, MapPost[]>,
-  region: string | null
-): boolean {
-  if (region !== null && getPlaceRegionSlug(place.area) !== region) return false;
-  if (!hasPostLevelFilter) return true;
-  return (matchedPostsByPlaceId.get(place.id)?.length ?? 0) > 0;
-}
-
-// 토픽 매칭은 항상 태그보다 위. 같은 급 안에서는 먼저 선택한 필터 기준.
-// 합산이 아닌 "가장 먼저 선택된 매칭"의 index만 사용해 선택 순서가 다중 매칭 누적에 묻히지 않도록.
-const TOPIC_BASE = 1000;
-const TAG_BASE = 1;
-function placeMatchScore(
-  place: Pick<MapPlace, "posts">,
-  topicIds: string[],
-  tagIds: string[]
-): number {
-  const bestTopicIdx = topicIds.findIndex((id) =>
-    place.posts.some((post) => post.topics.some((t) => topicMatchesFilter(t, id)))
-  );
-  if (bestTopicIdx !== -1) return TOPIC_BASE + (topicIds.length - bestTopicIdx);
-
-  const bestTagIdx = tagIds.findIndex((id) =>
-    place.posts.some((post) => post.tags.some((tag) => tag.id === id))
-  );
-  if (bestTagIdx !== -1) return TAG_BASE + (tagIds.length - bestTagIdx);
-
-  return 0;
-}
-
 type DiscoverSuggestion =
   | { type: "keyword"; text: string }
   | { type: "post"; text: string; placeName: string; placeId: string };
-
-/** topicChipMap과 동일한 chip 레벨 결정 로직으로 slug → id 반환 */
-function findTopicIdBySlug(topicTree: Level0TopicDeep[], slug: string): string | null {
-  for (const root of topicTree) {
-    const l1s = root.children;
-    if (l1s.length === 0) continue;
-    if (root.nameEn === KPOP_NAME) {
-      for (const l1 of l1s)
-        for (const l2 of l1.children)
-          if (l2.slug === slug) return l2.id;
-      continue;
-    }
-    const hasL2 = l1s.some((l1) => l1.children.length > 0);
-    if (!hasL2) {
-      for (const l1 of l1s)
-        if (l1.slug === slug) return l1.id;
-    } else {
-      for (const l1 of l1s)
-        for (const l2 of l1.children)
-          if (l2.slug === slug) return l2.id;
-    }
-  }
-  return null;
-}
-
 
 interface Props {
   allPlaces: (MapPlace & { isSaved?: boolean })[];
