+++ src/types/index.ts
@@ -15,9 +15,3 @@ export type SourcePlatform =
   | "TIKTOK"
   | "NEWS"
   | "OTHER";
-
-// 토픽 필터 공통 타입 (TopicFilterRow에서 re-export)
-export type { TopicBase, Level0Topic, Level1Topic, Level2Topic, Level3Topic } from "@/components/TopicFilterRow";
-
-// 태그 필터 공통 타입 (TagFilterRow에서 re-export)
-export type { TagForFilter, TagGroupForFilter } from "@/components/TagFilterRow";
